preliminary refactor
